add delete and update function

+++ src/main/java/kr/or/yonsai/domain/BoardVO.java
@@ -14,7 +14,7 @@ import lombok.Data;
 @Data
 public class BoardVO {
 
-	private Long bno;
+	private int bno;
 	private String title;
 	private String content;
 	private String writer;
